Remove `export` from `cloneECDb` and `cloneIModel`
