+++ apps/full-stack-tests/src/hierarchies/imodel-hierarchy-merging/InstancesNodes.test.ts
@@ -6,16 +6,10 @@
 import { collect } from "presentation-test-utilities";
 import { omit } from "@itwin/core-bentley";
 import { createMergedIModelHierarchyProvider } from "@itwin/presentation-hierarchies";
+import { createChangedDbs } from "../../ECDbUtils.js";
 import { initialize, terminate } from "../../IntegrationTests.js";
 import { NodeValidators, validateHierarchy, validateHierarchyLevel } from "../HierarchyValidation.js";
-import {
-  createChangedDbs,
-  createHierarchyDefinitionFactory,
-  createMergedHierarchyProvider,
-  importQSchema,
-  importXYZSchema,
-  pickAndTransform,
-} from "./HierarchiesMerging.js";
+import { createHierarchyDefinitionFactory, createMergedHierarchyProvider, importQSchema, importXYZSchema, pickAndTransform } from "./HierarchiesMerging.js";
 
 describe("Hierarchies", () => {
   before(async function () {
