+++ apps/full-stack-tests/src/hierarchies/imodel-hierarchy-merging/PropertyRangeGroupingNodes.test.ts
@@ -5,16 +5,10 @@
 /* eslint-disable @typescript-eslint/no-this-alias */
 
 import { omit } from "@itwin/core-bentley";
+import { createChangedDbs } from "../../ECDbUtils.js";
 import { initialize, terminate } from "../../IntegrationTests.js";
 import { NodeValidators, validateHierarchy } from "../HierarchyValidation.js";
-import {
-  createChangedDbs,
-  createHierarchyDefinitionFactory,
-  createMergedHierarchyProvider,
-  importQSchema,
-  importXYZSchema,
-  pickAndTransform,
-} from "./HierarchiesMerging.js";
+import { createHierarchyDefinitionFactory, createMergedHierarchyProvider, importQSchema, importXYZSchema, pickAndTransform } from "./HierarchiesMerging.js";
 
 describe("Hierarchies", () => {
   before(async function () {
